Fix hit box of scaled sprites (#254)
* Fix hit box of scaled sprites

We want to take into account not only the translation part of the global
transform, but also the scaling.  At the moment, rotations are still not
dealt with.

I am also not sure how performant the `to_scale_rotation_translation()`
method is.  (I think a different API could be faster, because we do not
really need the decomposition.)

* use scaled sprites in example

* Fix formatting

---------

Co-authored-by: Aevyrie <aevyrie@gmail.com>