Change log level from info to debug and trace (#262)
* Change log level from info to debug and trace

This implements https://github.com/aevyrie/bevy_mod_picking/issues/259.

Messages linked to the "Noisy" debug state are logged with `trace!()`.
The rest are logged with `debug!()`. Filtering on this _could_ replace
the Noisy/Normal state distinction entirely (perhaps leaving the
DebugPickingMode state _only_ for hiding or showing the cursor text
overlay) but I'll leave that possibility for the future. This just
makes a minimal change.

* Fix formatting

---------

Co-authored-by: Aevyrie <aevyrie@gmail.com>